backend/requirements.txt
cryptography
eth_account
fastapi
groq
langsmith
motor
pydantic
pydantic_core
python-dotenv
Telethon
uvicorn
web3
